requirements.txt
beautifulsoup4==4.11.1
blurhash==1.1.4
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==5.1.1
idna==3.3
Mastodon.py==1.5.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
